requirements.txt
bcrypt
flask
Flask-Login
Flask-WTF
pyjwt
pymongo
python-dotenv
werkzeug